+++ src/main/java/com/github/jrtex/tarot/model/Fold.java
@@ -1,7 +1,7 @@
-package com.github.jrtex.tarot;
+package com.github.jrtex.tarot.model;
 
-import java.util.List;
 import java.util.ArrayList;
+import java.util.List;
 
 public class Fold {
