+++ src/main/java/com/github/jrtex/tarot/model/TarotDeck.java
@@ -1,4 +1,4 @@
-package com.github.jrtex.tarot;
+package com.github.jrtex.tarot.model;
 
 import java.util.List;
