+++ src/main/java/com/github/jrtex/tarot/swing/ContractPanel.java
@@ -1,8 +1,10 @@
-package com.github.jrtex.tarot;
+package com.github.jrtex.tarot.swing;
 
 import java.awt.Color;
 
 import javax.swing.JLabel;
+
+import com.github.jrtex.tarot.model.Contract;
 
 @SuppressWarnings("serial")
 public class ContractPanel extends JLabel{
