+++ src/main/java/com/github/jrtex/tarot/swing/TarotListener.java
@@ -1,4 +1,4 @@
-package com.github.jrtex.tarot;
+package com.github.jrtex.tarot.swing;
 
 public interface TarotListener {
